Rename route files to be clearer
