+++ src/server.ts
@@ -2,8 +2,8 @@ import express from "express";
 import cors from "cors";
 import helmet from "helmet";
 import PGConnection from "./model/PGConnection";
-import peepsRoute from "./routes/peeps";
-import usersRoute from "./routes/users";
+import peepsRoute from "./routes/peepsRoute";
+import usersRoute from "./routes/usersRoute";
 
 PGConnection.open();
 const app = express();
